+++ src/tekore/_auth/expiring/decor.py
@@ -3,7 +3,7 @@ from collections.abc import Callable
 from httpx import codes
 
 from tekore._sender import Request, Response
-from tekore._sender.error import get_error
+from tekore._sender.error import RefreshTokenInvalid, get_error
 
 from .token import Token
 
@@ -13,15 +13,18 @@ def handle_errors(request: Request, response: Response) -> None:
     if not codes.is_error(response.status_code):
         return
 
+    error_cls = get_error(response.status_code)
     if codes.is_client_error(response.status_code):
-        error_str = f"{response.status_code} {response.content['error']}"
+        error = response.content["error"]
+        error_str = f"{response.status_code} {error}"
         description = response.content.get("error_description", None)
         if description is not None:
             error_str += ": " + description
+        if error == "invalid_grant":
+            error_cls = RefreshTokenInvalid
     else:
         error_str = "Unexpected error!"
 
-    error_cls = get_error(response.status_code)
     raise error_cls(error_str, request=request, response=response)